big red diff on index.js
yus.
Aside from the half-meme commit message, index.js has been cleaned up:
- Modularized event handlers.
- Event handlers are to be added in directory 'event_listeners'.
- Command loader also uses the same technique, moved to index.js

Test first! Can't test this locally because of hardcoded stuff.
