--- handler/command.js
@@ -1,32 +1,0 @@
-const { readdirSync } = require("fs");
-
-const ascii = require("ascii-table");
-
-let table = new ascii("commands");
-table.setHeading("Command", "Load status");
-
-module.exports = (bot) => {
-    readdirSync("./commands/").forEach(dir => {
-        const commands = readdirSync(`./commands/${dir}/`).filter(f => f.endsWith(".js"));
-
-        for (let file of commands) {
-            let pull = require(`../commands/${dir}/${file}`);
-
-            if (pull.name) {
-                bot.commands.set(pull.name, pull);
-                table.addRow(file, 'Loaded');
-            } else {
-                table.addRow(file, `Errors  -> missing a help.name, or help.name is not a string.`);
-                continue;
-            }
-           if (pull.aliases && Array.isArray(pull.aliases)) {
-                pull.aliases.forEach(alias => bot.aliases.set(alias, pull.name));
-             }
-         }
-    });
-
-    console.log(table.toString()); // Just log all existing files.
-};
-
-
-
